[Source] Moved BH initialization to Runtime.
