+++ include/cocaine/logging.hpp
@@ -28,6 +28,22 @@
 #include <blackhole/keyword.hpp>
 #include <blackhole/logger/wrapper.hpp>
 
+#define COCAINE_LOG(_log_, _level_, ...) \
+    if(auto _record_ = (_log_)->open_record(_level_)) \
+        ::blackhole::aux::logger::make_pusher(*(_log_), _record_, __VA_ARGS__)
+
+#define COCAINE_LOG_DEBUG(_log_, ...) \
+    COCAINE_LOG(_log_, ::cocaine::logging::debug, __VA_ARGS__)
+
+#define COCAINE_LOG_INFO(_log_, ...) \
+    COCAINE_LOG(_log_, ::cocaine::logging::info, __VA_ARGS__)
+
+#define COCAINE_LOG_WARNING(_log_, ...) \
+    COCAINE_LOG(_log_, ::cocaine::logging::warning, __VA_ARGS__)
+
+#define COCAINE_LOG_ERROR(_log_, ...) \
+    COCAINE_LOG(_log_, ::cocaine::logging::error, __VA_ARGS__)
+
 namespace cocaine { namespace logging {
 
 DECLARE_KEYWORD(source, std::string)
@@ -44,21 +60,5 @@ demangle() -> std::string {
 }
 
 }} // namespace cocaine::logging
-
-#define COCAINE_LOG(_log_, _level_, ...) \
-    if(auto _record_ = (_log_)->open_record(_level_)) \
-        ::blackhole::aux::logger::make_pusher(*(_log_), _record_, __VA_ARGS__)
-
-#define COCAINE_LOG_DEBUG(_log_, ...) \
-    COCAINE_LOG(_log_, ::cocaine::logging::debug, __VA_ARGS__)
-
-#define COCAINE_LOG_INFO(_log_, ...) \
-    COCAINE_LOG(_log_, ::cocaine::logging::info, __VA_ARGS__)
-
-#define COCAINE_LOG_WARNING(_log_, ...) \
-    COCAINE_LOG(_log_, ::cocaine::logging::warning, __VA_ARGS__)
-
-#define COCAINE_LOG_ERROR(_log_, ...) \
-    COCAINE_LOG(_log_, ::cocaine::logging::error, __VA_ARGS__)
 
 #endif
